Break out VertexReader, prepare VertexReader for CrossSIMD use in software transform
