+++ CMakeLists.txt
@@ -2049,6 +2049,7 @@ set(GPU_SOURCES
 	GPU/Common/VertexDecoderCommon.h
 	GPU/Common/VertexDecoderHandwritten.cpp
 	GPU/Common/VertexDecoderHandwritten.h
+	GPU/Common/VertexReader.h
 	GPU/Common/TransformCommon.cpp
 	GPU/Common/TransformCommon.h
 	GPU/Common/IndexGenerator.cpp
