+++ GPU/GPU.vcxproj.filters
@@ -264,6 +264,9 @@
     <ClInclude Include="GPUStateSIMDUtil.h">
       <Filter>Common</Filter>
     </ClInclude>
+    <ClInclude Include="Common\VertexReader.h">
+      <Filter>Common</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Math3D.cpp">
